add auth user

+++ src/components/molecules/Navbar/Navbar.jsx
@@ -6,12 +6,14 @@ import { Gap, Button, } from '../../atoms';
 import './navbar.scss';
 import Swal from "sweetalert2";
 import { useEffect } from 'react/cjs/react.development';
+import { useHistory } from 'react-router';
 
 const Navbar = () => {
 
+     const history = useHistory()
      const [isLogin, setIsLogin] = useState(false);
      const [isAuth, setIsAuth] = useState(false);
-     const [isLoading, setIsLoading] = useState(false);
+     const [authUser, setAuthUser] = useState(null);
 
      const urlAPI = process.env.REACT_APP_API_URL;
 
@@ -20,14 +22,15 @@ const Navbar = () => {
           const token = localStorage.getItem('token');
 
           if (token) {
-               setIsLoading(true);
                await fetch(urlAPI + `logout?token=${token}`, {
                     method: 'GET'
-               }).then(res => {
-                    localStorage.clear();
-                    setIsAuth(false);
-                    Swal.fire({ 'icon': 'success', 'title': 'Logout Success', 'showConfirmButton': true, 'confirmButtonColor': "#0F70B7" });
-                    setIsLoading(false);
+               }).then(() => {
+                    Swal.fire({ 'icon': 'success', 'title': 'Logout Success', 'showConfirmButton': true, 'confirmButtonColor': "#0F70B7" })
+                         .then(() => {
+                              localStorage.clear();
+                              setIsAuth(false);
+                              history.push('/');
+                         });
                }).catch(err => {
                     console.log(err);
                });
@@ -35,9 +38,29 @@ const Navbar = () => {
           }
      }
 
+     const fetchAuthUser = async () => {
+          const apiFetch = await fetch(urlAPI + `getuser?token=${localStorage.getItem('token')}`, {
+               method: 'GET',
+          }).catch(err => {
+               console.log(err);
+          });
+          const res = await apiFetch.json();
+          if (res.user) {
+               setAuthUser(res.user);
+               setIsAuth(true);
+          } else {
+               setIsAuth(false);
+               Swal.fire({ icon: 'error', title: "error", text: "Your session are expired. Please login again", confirmButtonColor: "#0F70B7" })
+                    .then(() => {
+                         history.push('/');
+                    });
+          }
+     }
+
      useEffect(() => {
           if (localStorage.getItem('token')) {
                setIsAuth(true);
+               fetchAuthUser();
           }
      }, []);
 
@@ -59,9 +82,21 @@ const Navbar = () => {
                                    <Gap width={30} height={10} />
                                    {
                                         isAuth ?
-                                             <Button type={1} onClick={logout} isLoading={isLoading}>
-                                                  <h6 className="fw-bold m-0 p-1">Logout</h6>
-                                             </Button>
+                                             <li class="nav-item dropdown">
+                                                  {
+                                                       authUser &&
+                                                       <Fragment>
+                                                            <a className="textBlue1 text-decoration-none dropdown-toggle" href="#" id="navbarDropdown" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+                                                                 {authUser.name}
+                                                            </a>
+                                                            <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
+                                                                 <Link to="/cart" className="text-decoration-none text-reset"><li className="dropdown-item">Cart</li></Link>
+                                                                 <Link to="/orders" className="text-decoration-none text-reset"><li className="dropdown-item">Orders</li></Link>
+                                                                 <li onClick={logout} className="dropdown-item" style={{ cursor: "pointer" }}>Logout</li>
+                                                            </ul>
+                                                       </Fragment>
+                                                  }
+                                             </li>
                                              :
                                              <Fragment>
                                                   <Button type={1} onClick={() => setIsLogin(true)} data-bs-toggle="modal" data-bs-target="#authModal">
@@ -84,7 +119,7 @@ const Navbar = () => {
                                    <Gap height={40} />
                                    {
                                         isLogin ?
-                                             <Login changeSection={() => setIsLogin(false)} setIsAuth={(value) => setIsAuth(value)} isAuth={isAuth} />
+                                             <Login changeSection={() => setIsLogin(false)} setIsAuth={(value) => setIsAuth(value)} isAuth={isAuth} setAuthUser={() => fetchAuthUser()} />
                                              :
                                              <Register changeSection={() => setIsLogin(true)} />
                                    }
